Moved position type to geometry
