+++ src/path.rs
@@ -1,8 +1,7 @@
 use std::f32::consts::PI;
 use std::slice;
 
-use crate::geometry::{self, Transform2D};
-use crate::position::Position;
+use crate::geometry::{self, Position, Transform2D};
 
 mod cache;
 pub use cache::{Convexity, PathCache};
